Move InstancedConfiguration to alluxio.conf
